fix up validation
