+++ packages/client-core/src/admin/common/variables/route.ts
@@ -1,0 +1,17 @@
+export interface RouteColumn {
+  id: 'project' | 'route' | 'active'
+  label: string
+  minWidth?: number
+  align?: 'right'
+}
+
+export const routeColumns: RouteColumn[] = [
+  { id: 'project', label: 'Project', minWidth: 65 },
+  { id: 'route', label: 'Route', minWidth: 65 },
+  {
+    id: 'active',
+    label: 'Active',
+    minWidth: 65,
+    align: 'right'
+  }
+]
