+++ packages/client-core/src/admin/components/Project/index.tsx
@@ -1,6 +1,6 @@
 import Cached from '@mui/icons-material/Cached'
-import CleaningServicesIcon from '@mui/icons-material/CleaningServices'
 import Cross from '@mui/icons-material/Cancel'
+import CleaningServicesIcon from '@mui/icons-material/CleaningServices'
 import Button from '@mui/material/Button'
 import Grid from '@mui/material/Grid'
 import Paper from '@mui/material/Paper'
@@ -52,26 +52,6 @@ const Projects = () => {
   }
 
   type Order = 'asc' | 'desc'
-
-  function getComparator<Key extends keyof any>(
-    order: Order,
-    orderBy: Key
-  ): (a: { [key in Key]: number | string }, b: { [key in Key]: number | string }) => number {
-    return order === 'desc'
-      ? (a, b) => descendingComparator(a, b, orderBy)
-      : (a, b) => -descendingComparator(a, b, orderBy)
-  }
-
-  function stableSort<T>(array: T[], comparator: (a: T, b: T) => number) {
-    const stabilizedThis = array.map((el, index) => [el, index] as [T, number])
-    stabilizedThis.sort((a, b) => {
-      const order = comparator(a[0], b[0])
-      if (order !== 0) return order
-      return a[1] - b[1]
-    })
-    const returned = stabilizedThis.map((el) => el[0])
-    return returned
-  }
 
   interface EnhancedTableProps {
     object: string
